Graphs: Fix strong_signals data collection
dump1090 produces stats.json, this is read to produce the database the
graphs are based on. strong_signals is a counter which logs every message
with an RSSI of bigger than -3dBFS.
Currently this message counter is unlike the other message counters read
from the last1min instead of the total bucket. This creates bad data as
the rrd data source type used assumes a constantly increasing counter,
which the last1min bucket is not. When the last1min strong_signals
bucket decreases the database logs unkown, when last1min strong_signals
increase it logs the increase and divides it by the elapsed time which
assuming a relatively constant number of strong_signals per minute can
greatly underestimate the actual value of strong_signals received.
Change the data collection to use the total strong_signals bucket like
the other message logs. The data collection will log the difference
between readings and create a correct rate.