Graphs: Change scaling of Messages >-3dBFS
Display Messages >-3dBFS per 10min instead of per hour. Due to the now
correct logging of those messages the red area corresponding to them can
get very big and make the rest of the graph hard to read.  Displaying
them like the total messages received was also considered but that tends
to make the red area become very small and hard to see.